Add feature that allows response bot on select categories
The feature allows staff or users with appropriate trust level
to enable response by the response bot on select categories.
This feature allows the respond bot reply new topics within
a select category. The feature is available to select when
the "allow response bot to reply all new topics" setting
is unchecked.